Phpstan coverage

+++ src/Command/PullIp2LocationLiteDatabaseCommand.php
@@ -128,16 +128,22 @@ readonly class PullIp2LocationLiteDatabaseCommand
 		}
 	}
 
-	private static function recursiveRemoveFolder(string $dir): void
+	private static function recursiveRemoveFolder(string $targetDir): void
 	{
-		foreach (scandir($dir) as $item) {
+		/** @var list<string>|false $dir */
+		$dir = scandir($targetDir);
+
+		if ($dir === false) {
+			throw new \RuntimeException('Failed to read target directory: ' . $targetDir);
+		}
+		foreach ($dir as $item) {
 			if ($item === '.' || $item === '..') {
 				continue;
 			}
-			$path = "$dir/$item";
+			$path = "$targetDir/$item";
 			is_dir($path) ? self::recursiveRemoveFolder($path) : unlink($path);
 		}
-		rmdir($dir);
+		rmdir($targetDir);
 	}
 
 }
